Web scraping dynamic websites
This is the python code I have thus far to scrape from a dynamic website using the Selenium and Beautiful Soup modules for Python.

The code only scrapes from the website in view (page 1). To scrape all pages, I will add code to loop through the pages in the archive. I decided to post what I have so far because most websites seem to be dynamic and it could be useful.

Note that I used Chrome. In order to use it, I had to download chromedriver from https://chromedriver.chromium.org/. If you use a different browser, you will need to download the specific webdriver.

References I used to write the code:
https://towardsdatascience.com/data-science-skills-web-scraping-javascript-using-python-97a29738353f
https://www.youtube.com/watch?v=vcnomT0CP0Y
https://www.youtube.com/watch?v=-yVNqaxejVg